pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["rifqoi <alfurqon.rifqi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.3"
sqlmodel = "^0.0.8"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
onnxruntime = "^1.13.1"
opencv-python-headless = "^4.6.0"
Pillow = "^9.3.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
